docs: better toctrees
This makes 'index' type toctrees list headings,
which makes those pages useful to get an overview.

For example the Tutorial index now shows all steps
inside the tutorial.

The IOC tutorials are also now numbered,
which gives a sense that it must be followed
in that order.

Another example, the "Options reference" index
can be used as an index of all NixOS options